rename file extensions
