+++ lib/functions.ts
