+++ pages/ssg.tsx
@@ -1,22 +1,21 @@
-import PropTypes from 'prop-types'
 import Layout from '@/components/global/Layout'
 import {fetcher} from '@/lib/functions'
 import Link from 'next/link'
 import {Info} from '@/components/blocks/Alerts'
 
-export default function ServerSideRendering(props) {
+export default function StaticGeneration(props) {
   return (
     <Layout
-      title="Server-side Rendering (SSR)"
-      description="A server-side rendered page with data from a REST-API."
+      title="Static Generation (SSG)"
+      description="A statically generated page with data from a REST-API."
     >
       <div className="container">
-        <h1>Server-side Rendering (SSR) Example</h1>
+        <h1>Static Generation (SSG) Example</h1>
 
         <Info>
           The content below is sourced from the WordPress REST-API.{' '}
-          <a href="https://nextjs.org/docs/basic-features/data-fetching#getserversideprops-server-side-rendering">
-            Learn more about SSR.
+          <a href="https://nextjs.org/docs/basic-features/data-fetching#getstaticprops-static-generation">
+            Learn more about SSG.
           </a>
         </Info>
 
@@ -42,11 +41,7 @@ export default function ServerSideRendering(props) {
  *
  * @see https://nextjs.org/docs/basic-features/data-fetching#getstaticprops-static-generation
  */
-export async function getServerSideProps() {
+export async function getStaticProps() {
   const data = await fetcher('https://nextjs.wpengine.com/wp-json/wp/v2/posts')
   return {props: {data}}
-}
-
-ServerSideRendering.propTypes = {
-  data: PropTypes.object
 }
